+++ ETrade/src/main/java/com/mthree/etrade/dao/UserDao.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserDao extends JpaRepository<User, Integer> {
+public interface UserDao extends JpaRepository<User, Long> {
     User findByEmail(String email);  // Find user by email
 }
-
